pyproject.toml
[build-system]
requires = ["poetry"]
build-backend = "poetry.masonry.api"

[tool.poetry]
authors = ["py <py@py>"]
description = ""
license = "MIT"
name = "eso-banana"
version = "0.0.1"

[[tool.poetry.packages]]
include = "banana"

[tool.poetry.scripts]
eso-banana-script = "banana:scripts.periodical_script"

[tool.poetry.dependencies]
requests = ""
PyYAML = ""
